Some little change

+++ HARDWARE/PWMDevice/PWMDevice.h
@@ -12,11 +12,11 @@
 
 #endif
 
-extern u32 Steering_Angle1;
-extern u32 Steering_Angle2;
+extern short Steering_Angle1;
+extern short Steering_Angle2;
 extern u32 LED_Brightness1;
 extern u32 LED_Brightness2;
 
-void Steering_AngleGain(u8 Steeringx,float Angle_Gain);
+void Steering_AngleGain(u8 Steeringx,short Angle_Gain);
 void LED_BrightnessGain(u8 LEDx,u32 Brightness_Gain);
